mod_impl unsigned->long

+++ executer/execute/execute_funcs_arithmetic.cpp
@@ -11,7 +11,7 @@ double sub_impl(double x, double y) { return x - y; }
 double mul_impl(double x, double y) { return x * y; }
 // TODO error checking
 double div_impl(double x, double y) { return x / y; }
-double mod_impl(double x, double y) { return ((unsigned)x) % ((unsigned)y); }
+double mod_impl(double x, double y) { return ((long)x) % ((long)y); }
 
 static std::unordered_map<vmopcode, std::function<double(double, double)>> arithm_funcs = {
     {add_vmiop, add_impl},
